berserk feature,replace smallint with int, copy information from game/tournament instead of passing slices, maps or pointers sqlc config emit_empty_slices

+++ api/internal/database/auth.sql.go
@@ -45,7 +45,7 @@ type GetSessionRow struct {
 	CreatedAt time.Time
 	ExpiresAt time.Time
 	Username  *string
-	Role      int16
+	Role      int32
 }
 
 func (q *Queries) GetSession(ctx context.Context, id string) (GetSessionRow, error) {
