+++ api/internal/tournament/player.go
@@ -5,11 +5,11 @@ import "github.com/notnil/chess"
 type Player struct {
 	ID              int32
 	IsActive        bool
-	Score           int32
-	Scores          []int16
+	Score           int
+	Scores          []int
 	Rating          float64
-	Streak          int32
-	Opponents       map[int32]int16
+	Streak          int
+	Opponents       map[int32]int
 	LastPlayedColor chess.Color
 }
 
@@ -18,9 +18,9 @@ func NewPlayer(id int32, rating float64, connected bool) *Player {
 		ID:        id,
 		IsActive:  true,
 		Score:     0,
-		Scores:    make([]int16, 0),
+		Scores:    make([]int, 0),
 		Rating:    rating,
 		Streak:    0,
-		Opponents: make(map[int32]int16),
+		Opponents: make(map[int32]int),
 	}
 }
